requirements.txt
# Import the corresponding specific SQL RDBMS that we will be using as well, not just SQLAlchemy
# In our case, it will be PyMySQL

slackers==0.4.0
slackclient==2.2.1
sqlalchemy>=1.3.10
PyMySQL==0.9.3
google-cloud-firestore==1.6.1
pyee>=6.0.0
python-multipart>=0.0.5
requests>=2.22
uvloop==0.14.0
uvicorn==0.11.1
starlette==0.12.9
fastapi==0.47.1
certifi
